requirement.txt
streamlit
zai-sdk>=0.2.2
python-dotenv